Improve Java version detection
Java version detection of the Un*x start scripts depends on the version to be
in the first line and the implementation name in the second line. However this
is not always true, e.g. when properties are set via environment variable.

Example:
	export _JAVA_OPTIONS=""
	java -version

This commit will make the version position independent, however it still
assumes that the line containing the implementation name will follow the
version line.