Fix import of MS Project tasks without start (#119)
Don't fail if a task in an MS Project file does not explicitly
have a start time set.
Tasks do not necessarily have to have start time set,
e.g. the start of parent tasks may be derived from their child
tasks.

The fact that the start element is not mandatory
for each task is also described here:
https://msdn.microsoft.com/en-us/library/office/bb968645%28v=office.14%29.aspx